requirements.txt
afdko>=4.0.1
fonttools>=4.59.2
PyYAML>=6.0.2
pathos>=0.3.4
skia-pathops>=0.8.0